EES-4770 Move and rename ReleaseSeriesItem view model/request
